+++ sniffer/network/engine.py
@@ -1,6 +1,6 @@
 import socket
-from pwn import hexdump
-from ..utils.constants import ETH_P_ALL
+import typing
+from ..utils import constants
 from .analyzer import PacketAnalyzer
 from ..utils.decorators import require_root
 from ..exceptions.network import UninterestingPacketException,\
@@ -20,7 +20,7 @@ class SnifferEngine:
 
         self.socket = socket.socket(socket.AF_PACKET,
                                     socket.SOCK_RAW,
-                                    socket.ntohs(ETH_P_ALL))
+                                    socket.ntohs(constants.ETH_P_ALL))
 
         # self.socket.setsockopt(socket.IPPROTO_IP, socket.IP_HDRINCL, 1)
         # self.socket.ioctl(socket.SIO_RCVALL, socket.RCVALL_ON)
@@ -32,35 +32,39 @@ class SnifferEngine:
         # self.socket.setsockopt(socket.IPPROTO_IP, socket.IP_HDRINCL, 1)
 
     @property
-    def total_packet_count(self):
+    def total_packet_count(self) -> int:
         return self._packet_count
 
     @total_packet_count.setter
-    def total_packet_count(self, value):
+    def total_packet_count(self, value) -> None:
         self._packet_count = value
 
     @property
-    def http_packet_count(self):
+    def http_packet_count(self) -> int:
         return self._http_packet_count
 
     @http_packet_count.setter
-    def http_packet_count(self, value):
+    def http_packet_count(self, value) -> None:
         self._http_packet_count = value
 
-    async def __sniff(self):
+    async def __sniff(self) -> typing.AsyncGenerator[bytes, None]:
         while True:
             yield self.socket.recvfrom(SnifferEngine.MAX_PACKET_LEN)[0]
 
-    async def sniff(self, count: int = -1):
+    async def sniff(self,
+                    count: int) -> typing.AsyncGenerator[PacketAnalyzer, None]:
         async for packet in self.__sniff():
+            if count != constants.INFINITY and self.http_packet_count >= count:
+                break
+
             self.total_packet_count += 1
 
             try:
                 analyzer = PacketAnalyzer(packet, self.http_packet_count)
-                # print(hexdump(packet))
                 self.http_packet_count += 1
 
                 yield self.http_packet_count, analyzer
+            # Skip over uninteresting packets
             except UnsupportedVersionException:
                 continue
             except UninterestingPacketException: